TIP-860: Move CategoryNormalizer for Versioning from Tool to Pim/Enrichment
