+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/serializers_versionning.yml
@@ -1,0 +1,12 @@
+parameters:
+    pim_serializer.normalizer.flat.category.class: Akeneo\Pim\Enrichment\Component\Category\Normalizer\Versioning\CategoryNormalizer
+
+
+services:
+    pim_versioning.serializer.normalizer.flat.category:
+        class: '%pim_serializer.normalizer.flat.category.class%'
+        arguments:
+            - '@pim_catalog.normalizer.standard.category'
+            - '@pim_versioning.serializer.normalizer.flat.label_translation'
+        tags:
+            - { name: pim_versioning.serializer.normalizer, priority: 90 }
